requirements.txt
alembic
apscheduler
cryptography
fastapi
faster-whisper
freezegun
httpx
orjson
passlib
pydantic-settings
pyjwt
pytest
pytest-asyncio
python-dateutil
python-multipart
requests
sqlalchemy>=2
tenacity
uvicorn
wechatpy
